Fix a number of problems with ELCAssetTablePicker: - It continued loading an album until finished, even if it had been dismissed/dealloced - It preloaded 2 UIImageViews per asset: slow plus memory problems w/ large albums - It contained unsafe_unretained references to its parent and ALAssetGroup, causing crashes if it was dismissed while loading, etc
This commit rewrites ELCAsset and ELCAssetCell to make view creation the responsibility of ELCAssetCell. That's how table views are supposed to work.
This sacrifices some scrolling performance. To mitigate that, a UIImage for each asset's thumb is cached in ELCAsset after it is first needed. These can be purged to save memory when a memory warning is received.